added styling to the about page
